+++ fltk-sys/cfltk/cfl_browser.h
@@ -46,7 +46,9 @@ extern "C" {
     void widget##_set_scrollbar_size(widget *self, int newSize);                                   \
     int widget##_scrollbar_width(const widget *self);                                              \
     void widget##_set_scrollbar_width(widget *self, int width);                                    \
-    void widget##_sort(widget *self);
+    void widget##_sort(widget *self);                                                              \
+    void *widget##_scrollbar(widget *self);                                                        \
+    void *widget##_hscrollbar(widget *self);
 
 WIDGET_DECLARE(Fl_Browser)